requirements.txt
asgiref==3.7.2
Django==4.2.2
fabric==2.6.0
djangorestframework==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-ckeditor==6.5.1
python-dotenv==1.0.0
pytz==2023.3
psycopg2==2.9.6
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
gunicorn==20.1.0
whitenoise==6.5.0
Pillow==9.5.0
PyJWT==2.7.0